// src/units/types/common/UnitStat.ts
/** ステータス */
export type UnitStat = {
  /** HP */
  hp: number;
  /** MP */
  mp: number;
  /** 攻撃 */
  atk: number;
  /** 防御 */
  def: number;
  /** 魔力 */
  matk: number;
  /** 魔防 */
  mdef: number;
  /** 腕力 */
  power: number;
  /** 耐久 */
  endure: number;
  /** 幸運 */
  luck: number;
  /** 知性 */
  intelligence: number;
  /** 精神 */
  split: number;
  /** 速度 */
  speed: number;
};